renaming; pom file adjustments
