+++ pass-grant-cli/pom.xml
@@ -43,6 +43,7 @@
                         </goals>
                         <configuration>
                             <shadedArtifactAttached>true</shadedArtifactAttached>
+                            <finalName>jhu-grant-loader-${jhu.loader.version}</finalName>
                             <transformers>
                                 <transformer implementation=
                                                      "org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
