+++ pom.xml
@@ -73,6 +73,8 @@
         <slf4j.version>1.7.25</slf4j.version>
         <unitils.version>3.4.6</unitils.version>
 
+        <jhu.loader.version>1.0.0</jhu.loader.version>
+
     </properties>
 
     <build>
